backend/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
asgiref==3.11.1
asttokens==3.0.0
attrs==26.1.0
blinker==1.9.0
build==1.4.2
certifi==2026.2.25
cfgv==3.4.0
charset-normalizer==3.4.5
click==8.3.0
colorama==0.4.6
comm==0.2.2
contourpy==1.3.1
cycler==0.12.1
debugpy==1.8.12
decorator==5.2.1
distlib==0.4.0
Django==6.0.4
django-cors-headers==4.9.0
djangorestframework==3.17.1
djangorestframework_simplejwt==5.5.1
docutils==0.22.4
drf-spectacular==0.29.0
executing==2.2.0
filelock==3.19.1
flake8==7.3.0
Flask==3.1.2
fonttools==4.55.3
frozenlist==1.8.0
id==1.6.1
identify==2.6.13
idna==3.11
ImageHash==4.3.2
inflection==0.5.1
iniconfig==2.3.0
ipykernel==6.29.5
ipython==8.32.0
itsdangerous==2.2.0
jaraco.classes==3.4.0
jaraco.context==6.1.2
jaraco.functools==4.4.0
jedi==0.19.2
Jinja2==3.1.6
joblib==1.4.2
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
jupyter_client==8.6.3
jupyter_core==5.7.2
keyring==25.7.0
kiwisolver==1.4.8
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.0
matplotlib-inline==0.1.7
mccabe==0.7.0
mdurl==0.1.2
more-itertools==11.0.1
multidict==6.7.1
nest-asyncio==1.6.0
nh3==0.3.4
nodeenv==1.9.1
numpy==2.2.2
packaging==24.2
pandas==2.2.3
parso==0.8.4
pillow==11.1.0
platformdirs==4.3.6
pluggy==1.6.0
pre_commit==4.3.0
prompt_toolkit==3.0.50
propcache==0.4.1
psutil==7.0.0
psycopg2==2.9.9
pure_eval==0.2.3
pycodestyle==2.14.0
pyflakes==3.4.0
pygame==2.5.2
Pygments==2.19.1
PyJWT==2.12.1
pyparsing==3.2.1
pyproject_hooks==1.2.0
pytest==8.4.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
pytz==2024.2
PyWavelets==1.9.0
pywin32==308
pywin32-ctypes==0.2.3
PyYAML==6.0.2
pyzmq==26.2.1
readme_renderer==44.0
referencing==0.37.0
requests==2.32.5
requests-toolbelt==1.0.0
rfc3986==2.0.0
rich==14.3.3
rpds-py==0.30.0
scikit-learn==1.6.1
scipy==1.15.1
seaborn==0.13.2
six==1.17.0
sqlparse==0.5.5
stack-data==0.6.3
threadpoolctl==3.5.0
tornado==6.4.2
traitlets==5.14.3
twine==6.2.0
typing_extensions==4.15.0
tzdata==2026.1
uritemplate==4.2.0
urllib3==2.6.3
virtualenv==20.34.0
wcwidth==0.2.13
Werkzeug==3.1.3
yarl==1.23.0
